Moved network dependency to a utils location, eliminating duplicate code.
